requirements.txt
asgiref==3.11.0
dj-database-url==2.2.0
Django==4.2.11
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
boto3==1.35.76
django-cors-headers==4.6.0
Pillow==11.0.0